Remove GitHub test workflow and update dependencies
This commit removes the test.yml GitHub workflow and updates dependencies in Cargo.toml. Additionally, it includes Clippy warnings at the start of main.rs and increments the package version. It also adds the ci.yml GitHub workflow. These changes ensure the latest functionality from dependencies is utilized and a CI workflow is in place for checks on pull requests and pushes on the master branch. Clippy warnings will enforce best coding practices for maintainability.