Update dependencies and improve Docker setup
Upgraded various dependencies, including changing 'base64' from version '0.20.0' to '0.21.4', and a number of 'kube' packages from version '0.86.0' to '0.87.1'. In addition, several changes were made to the Docker setup and CI pipeline, such as enabling cross-compiling in the Docker image workflow and removing redundant parts of the 'build' command in the CI workflow. Minor cleanups were also done such as replacing 'http' with 'axum::http' in two files for consistency.